agent/pyproject.toml
[project]
name = "quorix"
version = "0.1.0"
description = "Your ever-curious, voice-powered thinking companion that listens, remembers, and helps you untangle ideas like a genius rubber duck with a brain upgrade."
authors = [
    {name = "Muddassir Lateef",email = "lateef.muddassir@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "livekit-agents[groq] (>=1.0,<2.0)",
    "livekit-plugins-silero (>=1.0.20,<2.0.0)",
    "livekit-plugins-groq (>=1.0.20,<2.0.0)",
    "livekit-plugins-turn-detector (>=1.0.20,<2.0.0)",
    "livekit-plugins-noise-cancellation (>=0.2.3,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
